+++ src/wam2layers/cli.py
@@ -17,7 +17,8 @@ import click
 from wam2layers.analysis import visualization
 from wam2layers.config import Config
 from wam2layers.preprocessing.era5 import prep_experiment
-from wam2layers.tracking.backtrack import run_experiment
+from wam2layers.tracking.backtrack import run_experiment as run_backtrack_experiment
+from wam2layers.tracking.backtrack import run_experiment as run_forwardtrack_experiment
 
 logger = logging.getLogger(__name__)
 
@@ -60,14 +61,13 @@ def cli():
 @cli.command()
 @click.argument("config_file", type=click.Path(exists=True))
 def backtrack(config_file):
-    """Run WAM2layers backtrack experiment from the command line.
+    """Run WAM2layers backtrack experiment.
 
     CONFIG_FILE: Path to WAM2layers experiment configuration file.
 
     Usage examples:
 
         \b
-        - python path/to/backtrack.py path/to/cases/era5_2021.yaml
         - wam2layers backtrack path/to/cases/era5_2021.yaml
     """
     log_path = Config.from_yaml(config_file).output_folder
@@ -75,12 +75,33 @@ def backtrack(config_file):
     _copy_config_yaml(config_file, log_path)
     logger.info("Welcome to WAM2layers.")
     logger.info("Starting backtrack experiment.")
-    run_experiment(config_file)
+    run_backtrack_experiment(config_file)
+
+
+# Command line setup for forwardtrack
+
+
+@cli.command()
+@click.argument("config_file", type=click.Path(exists=True))
+def forwardtrack(config_file):
+    """Run WAM2layers forwardtrack experiment.
+
+    CONFIG_FILE: Path to WAM2layers experiment configuration file.
+
+    Usage examples:
+
+        \b
+        - wam2layers forwardtrack path/to/cases/era5_2021.yaml
+    """
+    log_path = Config.from_yaml(config_file).output_folder
+    setup_logging(log_path)
+    _copy_config_yaml(config_file, log_path)
+    logger.info("Welcome to WAM2layers.")
+    logger.info("Starting forwardtrack experiment.")
+    run_forwardtrack_experiment(config_file)
 
 
 # Command line setup for preprocess
-
-
 @click.group()
 def preproc_cli():
     """Pre-process raw input data for tracking with WAM2layers"""
@@ -97,7 +118,6 @@ def era5(config_file):
     Usage examples:
 
         \b
-        - python path/to/preprocessing/era5.py path/to/cases/era5_2021.yaml
         - wam2layers preprocess era5 path/to/cases/era5_2021.yaml
     """
     log_path = Config.from_yaml(config_file).preprocessed_data_folder
@@ -109,8 +129,6 @@ def era5(config_file):
 
 
 # Command line setup for visualization
-
-
 @click.group()
 def visualize_cli():
     """Visualize input or output data of a WAM2layers experiment"""