+++ src/wam2layers/tracking/forwardtrack.py
@@ -23,7 +23,7 @@ from wam2layers.utils.profiling import ProgressTracker
 logger = logging.getLogger(__name__)
 
 
-def backtrack(
+def forwardtrack(
     F,
     S1,
     S0,
@@ -39,62 +39,84 @@ def backtrack(
     evap = F["evap"].values
     precip = F["precip"].values
     f_vert = F["f_vert"].values
-    s_upper = S1["s_upper"].values
-    s_lower = S1["s_lower"].values
+    s_upper = S0["s_upper"].values
+    s_lower = S0["s_lower"].values
 
     s_track_upper = output["s_track_upper_restart"].values
     s_track_lower = output["s_track_lower_restart"].values
 
     # Short name for often used expressions
+    precip_lower = precip * s_lower / (s_upper + s_lower)
+    precip_upper = precip * s_upper / (s_upper + s_lower)
     s_track_relative_lower = np.minimum(s_track_lower / s_lower, 1.0)
     s_track_relative_upper = np.minimum(s_track_upper / s_upper, 1.0)
 
-    # Actual tracking (note: backtracking, all fluxes change sign)
     bc = config.periodic_boundary  # boundary condition True/False
     # TODO: apply terms in successive steps instead of all at once?
     s_track_lower += (
-        +horizontal_advection(s_track_relative_lower, -fx_lower, -fy_lower, bc)
-        + vertical_advection(-f_vert, s_track_relative_lower, s_track_relative_upper)
+        horizontal_advection(s_track_relative_lower, fx_lower, fy_lower, bc)
+        + vertical_advection(f_vert, s_track_relative_lower, s_track_relative_upper)
         + vertical_dispersion(
-            -f_vert, s_track_relative_lower, s_track_relative_upper, config.kvf
+            f_vert, s_track_relative_lower, s_track_relative_upper, config.kvf
         )
-        + region * precip * s_lower / (s_upper + s_lower)
-        - evap * s_track_relative_lower
+        + region * evap
+        - precip_lower * s_track_relative_lower
     )
+    # TODO: find better way to deal with negative values
+    s_track_negative_lower = s_track_lower[s_track_lower < 0] / S1["s_lower"]
+    if np.any(s_track_negative_lower):
+        logger.warn(
+            f"""Negative values encountered in s_track_lower. Count, minimum:
+                    {s_track_negative_lower.count(), s_track_negative_lower.minimum()}"""
+        )
+    s_track_lower = np.minimum(s_track_lower, 0)
 
     s_track_upper += (
-        +horizontal_advection(s_track_relative_upper, -fx_upper, -fy_upper, bc)
-        - vertical_advection(-f_vert, s_track_relative_lower, s_track_relative_upper)
+        horizontal_advection(s_track_relative_upper, fx_upper, fy_upper, bc)
+        - vertical_advection(f_vert, s_track_relative_lower, s_track_relative_upper)
         - vertical_dispersion(
-            -f_vert, s_track_relative_lower, s_track_relative_upper, config.kvf
+            f_vert, s_track_relative_lower, s_track_relative_upper, config.kvf
         )
-        + region * precip * s_upper / (s_upper + s_lower)
+        - precip_upper * s_track_relative_upper
     )
+    # TODO: find better way to deal with negative values
+    s_track_negative_upper = s_track_upper[s_track_upper < 0] / S1["s_upper"]
+    if np.any(s_track_negative_upper):
+        logger.warn(
+            f"""Negative values encountered in s_track_upper. Count, minimum:
+                    {s_track_negative_upper.count(), s_track_negative_upper.minimum()}"""
+        )
+    s_track_upper = np.minimum(s_track_upper, 0)
 
     # down and top: redistribute unaccounted water that is otherwise lost from the sytem
     # TODO build in logging for lost moisture
-    lower_to_upper = np.maximum(0, s_track_lower - S0["s_lower"])
-    upper_to_lower = np.maximum(0, s_track_upper - S0["s_upper"])
+    overshoot_lower = np.maximum(0, s_track_lower - S1["s_lower"])
+    overshoot_upper = np.maximum(0, s_track_upper - S1["s_upper"])
     s_track_lower = np.minimum(
-        s_track_lower - lower_to_upper + upper_to_lower, S0["s_lower"]
+        s_track_lower - overshoot_lower + overshoot_upper, S1["s_lower"]
     )
     s_track_upper = np.minimum(
-        s_track_upper - upper_to_lower + lower_to_upper, S0["s_upper"]
+        s_track_upper - overshoot_upper + overshoot_lower, S1["s_upper"]
     )
 
     # Update output fields
-    output["e_track"] += evap * s_track_relative_lower
+    output["p_track_lower"] += precip_lower * s_track_relative_lower
+    output["p_track_upper"] += precip_upper * s_track_relative_upper
 
     # Bookkeep boundary losses as "tracked moisture at grid edges"
-    output["e_track"][0, :] += (s_track_upper + s_track_lower)[0, :]
-    output["e_track"][-1, :] += (s_track_upper + s_track_lower)[-1, :]
+    output["p_track_lower"][0, :] += s_track_lower[0, :]
+    output["p_track_lower"][-1, :] += s_track_lower[-1, :]
+    output["p_track_upper"][0, :] += s_track_upper[0, :]
+    output["p_track_upper"][-1, :] += s_track_upper[-1, :]
     s_track_upper[0, :] = 0
     s_track_upper[-1, :] = 0
     s_track_lower[0, :] = 0
     s_track_lower[-1, :] = 0
     if config.periodic_boundary is False:  # bookkeep west and east losses
-        output["e_track"][:, 0] += (s_track_upper + s_track_lower)[:, 0]
-        output["e_track"][:, -1] += (s_track_upper + s_track_lower)[:, -1]
+        output["p_track_lower"][:, 0] += s_track_lower[:, 0]
+        output["p_track_lower"][:, -1] += s_track_lower[:, -1]
+        output["p_track_upper"][:, 0] += s_track_upper[:, 0]
+        output["p_track_upper"][:, -1] += s_track_upper[:, -1]
         s_track_upper[:, 0] = 0
         s_track_upper[:, -1] = 0
         s_track_lower[:, 0] = 0
@@ -102,7 +124,7 @@ def backtrack(
 
     output["s_track_lower_restart"].values = s_track_lower
     output["s_track_upper_restart"].values = s_track_upper
-    output["tagged_precip"] += region * precip
+    output["tagged_evap"] += region * evap
 
 
 def initialize(config_file):
@@ -135,8 +157,9 @@ def initialize_outputs(region):
             # Keep last state for a restart
             "s_track_upper_restart": xr.zeros_like(proto),
             "s_track_lower_restart": xr.zeros_like(proto),
-            "e_track": xr.zeros_like(proto),
-            "tagged_precip": xr.zeros_like(proto),
+            "p_track_upper": xr.zeros_like(proto),
+            "p_track_lower": xr.zeros_like(proto),
+            "tagged_evap": xr.zeros_like(proto),
         }
     )
     return output
@@ -166,15 +189,15 @@ def initialize_time(config, direction="forward"):
 
 
 def run_experiment(config_file):
-    """Run a backtracking experiment from start to finish."""
+    """Run a forwardtracking experiment from start to finish."""
     config, region, output = initialize(config_file)
 
     event_start, event_end = config.event_start_date, config.event_end_date
-    progress_tracker = ProgressTracker(output, mode="backtrack")
+    progress_tracker = ProgressTracker(output, mode="forwardtrack")
 
-    t0, th, t1, dt = initialize_time(config, direction="backward")
+    t0, th, t1, dt = initialize_time(config, direction="forward")
 
-    while t0 >= config.track_start_date:
+    while t1 <= config.track_end_date:
         S0 = load_data(t0, config, "states")
         F = load_data(th, config, "fluxes")
         S1 = load_data(t1, config, "states")
@@ -185,29 +208,30 @@ def run_experiment(config_file):
         change_units(S1, config.target_frequency)
 
         # Apply a stability correction if needed
-        stabilize_fluxes(F, S1, progress_tracker, config, t1)
+        stabilize_fluxes(F, S0, progress_tracker, config, t0)
 
         # Determine the vertical moisture flux
         F["f_vert"] = calculate_fz(F, S0, S1, config.kvf)
 
         # Only track the precipitation at certain timesteps
-        if not event_start <= t1 <= event_end:
-            F["precip"] = 0
+        if not event_start <= t0 <= event_end:
+            F["evap"] = 0
 
-        # Inside backtrack the "output" dictionary is updated
-        backtrack(F, S1, S0, region, output, config)
-        t1 -= dt
-        th -= dt
-        t0 -= dt
+        # TODO: consider changing signature to F, S0, S1
+        # Inside forwardtrack the "output" dictionary is updated
+        forwardtrack(F, S1, S0, region, output, config)
+        t1 += dt
+        th += dt
+        t0 += dt
 
         # Daily output
-        is_output_time = t1 == t1.floor(config.output_frequency)
-        is_final_step = t0 < config.track_start_date
+        is_output_time = t0 == t0.floor(config.output_frequency)
+        is_final_step = t1 > config.track_end_date
         if is_output_time or is_final_step:
-            progress_tracker.print_progress(t1, output)
+            progress_tracker.print_progress(t0, output)
             progress_tracker.store_intermediate_states(output)
-            write_output(output, t1, config, mode="backtrack")
+            write_output(output, t0, config, mode="forwardtrack")
             # Flush previous outputs
-            output[["e_track", "tagged_precip"]] *= 0
+            output[["p_track_upper", "p_track_lower", "tagged_evap"]] *= 0
 
     logger.info("Experiment complete.")