+++ src/wam2layers/utils/profiling.py
@@ -28,8 +28,9 @@ class Profiler:
 
 
 class ProgressTracker:
-    def __init__(self, output):
+    def __init__(self, output, mode="backtrack"):
         """Keep track of tagged and tracked moisture."""
+        self.mode = mode
         self.total_tagged_moisture = 0
         self.tracked = 0
         self.store_intermediate_states(output)
@@ -44,27 +45,39 @@ class ProgressTracker:
         don't lose accumulations from previous outputs.
         """
         totals = output.sum()
-        self.tracked += totals["e_track"]
-        self.total_tagged_moisture += totals["tagged_precip"]
+        if self.mode == "backtrack":
+            self.tracked += totals["e_track"]
+            self.total_tagged_moisture += totals["tagged_precip"]
+        else:  # mode is forwardtrack
+            self.tracked += totals["p_track_lower"] + totals["p_track_upper"]
+            self.total_tagged_moisture += totals["tagged_evap"]
 
     def print_progress(self, t, output):
         """Print some useful runtime diagnostics."""
         totals = output.sum()
-        tracked = self.tracked + totals["e_track"]
-        total_tagged_moisture = self.total_tagged_moisture + totals["tagged_precip"]
+
+        if self.mode == "backtrack":
+            tracked = self.tracked + totals["e_track"]
+            total_tagged_moisture = self.total_tagged_moisture + totals["tagged_precip"]
+        else:  # mode is forwardtrack
+            tracked = self.tracked + totals["p_track_upper"] + totals["p_track_lower"]
+            total_tagged_moisture = self.total_tagged_moisture + totals["tagged_evap"]
         still_in_atmosphere = (
             totals["s_track_upper_restart"] + totals["s_track_lower_restart"]
         )
-
-        total_tracked_moisture = tracked + still_in_atmosphere
+        # total_tracked_moisture = tracked + still_in_atmosphere
+        # tracked_percentage = tracked / total_tagged_moisture * 100
+        # lost_percentage = (1 - total_tracked_moisture / total_tagged_moisture) * 100
         tracked_percentage = tracked / total_tagged_moisture * 100
-        lost_percentage = (1 - total_tracked_moisture / total_tagged_moisture) * 100
+        in_atmos_percentage = still_in_atmosphere / total_tagged_moisture * 100
+        lost_percentage = 100 - tracked_percentage - in_atmos_percentage
 
         time, memory = self.profile()
 
         logger.info(
             f"{t} - "
             f"Tracked moisture: {tracked_percentage.item():.2f}%. "
+            f"Moisture in atmosphere: {in_atmos_percentage.item():.2f}%. "
             f"Lost moisture: {lost_percentage.item():.2f}%. "
             f"Time since start: {time}s, RAM: {memory:.2f} MB"
         )