+++ CHANGELOG.md
@@ -6,11 +6,13 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Add forward tracking possibility including small bugfix in backtrack ([#289](https://github.com/WAM2layers/WAM2layers/pull/289)).
 - Add tests for preprocess and backtrack workflow ([#265](https://github.com/WAM2layers/WAM2layers/pull/265)).
 - Add tests for visualize workflow ([#271](https://github.com/WAM2layers/WAM2layers/pull/271)).
 - Add `pre-commit` to check and fix code with `black` and `isort` formatter ([#273](https://github.com/WAM2layers/WAM2layers/pull/273)).
 - Add additional documentation and doctest for vertical advection terms ([#274](https://github.com/WAM2layers/WAM2layers/pull/274)).
 - Copy config yaml file to output path ([#249](https://github.com/WAM2layers/WAM2layers/pull/249)).
+- Add forward tracking ([#289](https://github.com/WAM2layers/WAM2layers/pull/289))
 
 ### Removed